refactor: 도서관의 재고 검색 로직을 재고 패키지로 이동
